{{cookiecutter.project_slug}}/requirements_dev.txt
# Development requirements for {{cookiecutter.module_name}}. True production requirements go in setup.py
pytest
invoke
sphinx
OpenEye-floe-pkg-tools
OpenEye-Artemis
sphinx_rtd_theme

-e .
